+++ docs/_plugins/rhds.ts
@@ -1,9 +1,9 @@
 import type { UserConfig } from '@11ty/eleventy';
-import * as ChildProcess from 'node:child_process';
-import { join, dirname } from 'node:path';
-import { promisify } from 'node:util';
-import { readdir, readFile, writeFile, mkdir } from 'node:fs/promises';
+import { join, dirname, relative } from 'node:path';
+import { cp, glob, readdir, readFile, writeFile, mkdir } from 'node:fs/promises';
 import { makeDemoEnv } from '#scripts/environment.js';
+
+import { $ } from 'execa';
 
 import yaml from 'js-yaml';
 import slugify from 'slugify';
@@ -19,9 +19,6 @@ const repoStatus = yaml.load(await readFile(
   new URL('../_data/repoStatus.yaml', import.meta.url),
   'utf8',
 ));
-
-const exec = promisify(ChildProcess.exec);
-const cwd = process.cwd();
 
 /**
  * EleventyTransformContext the `this` binding for transform functions
@@ -59,6 +56,8 @@ const COPY_CONTENT_EXTENSIONS = [
   'map',
   'd.ts',
 ];
+
+const cwd = process.cwd();
 
 /**
  * Generate a map of files per package which should be copied to the site dir
@@ -108,7 +107,7 @@ export default async function(eleventyConfig: UserConfig, options?: Options) {
   /** custom-elements.json */
   eleventyConfig.on('eleventy.before', async function({ runMode }) {
     if (runMode === 'watch') {
-      await exec('npx cem analyze');
+      await $`npx cem analyze`;
     }
   });
 
@@ -138,6 +137,43 @@ export default async function(eleventyConfig: UserConfig, options?: Options) {
 
   eleventyConfig.addPassthroughCopy(await getFilesToCopy(), {
     filter: (path: string) => !path.endsWith('.html'),
+  });
+
+  eleventyConfig.on('eleventy.after', async function({ runMode }) {
+    const cwd = process.cwd();
+    const pkgsDir = join(cwd, '_site/assets/packages');
+    switch (runMode) {
+      case 'build':
+        await $`npx tspc -b elements`;
+        await mkdir(join(pkgsDir, '@rhds/elements/elements'), { recursive: true });
+        await mkdir(join(pkgsDir, '@rhds/elements/lib'), { recursive: true });
+        for await (const file of glob('./{elements,lib}/**/*.{js,d.ts,map,css}')) {
+          const outDir = join(pkgsDir, '@rhds/elements');
+          const rel = relative(cwd, file);
+          const out = join(outDir, dirname(rel));
+          const from = join(cwd, rel);
+          const to = join(outDir, rel);
+          await mkdir(out, { recursive: true });
+          await cp(from, to);
+        }
+    }
+  });
+
+  eleventyConfig.on('eleventy.after', async function({ runMode }) {
+    const cwd = process.cwd();
+    const pkgsDir = join(cwd, '_site/assets/packages');
+    switch (runMode) {
+      case 'build':
+        await $`npx tspc -b elements`;
+        await mkdir(join(pkgsDir, '@uxdot/elements'), { recursive: true });
+        for await (const file of glob('./uxdot/*.{js,d.ts,map,css}')) {
+          const rel = relative(join(cwd, 'uxdot'), file);
+          const outDir = join(pkgsDir, '@uxdot/elements');
+          const from = join(cwd, 'uxdot', rel);
+          const to = join(outDir, rel);
+          await cp(from, to);
+        }
+    }
   });
 
   eleventyConfig.addJavaScriptFunction('getTagNameSlug', getTagNameSlug);
